Quotient and remainder loops added.
It works for the tested polynomials, but higher divisor ones need to be looked at more. There's also several niche cases that need to be handled to prevent users from entering bad polys, like a zero polynomial, one where the power of the divisor is greater than dividend, and divisor polys with trailing zeroes that'll cause divide by 0 errors.